API/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
pydantic>=2.8,<3
pydantic-settings>=2.3,<3
PyJWT>=2.8,<3
cachetools>=5.3,<6

